+++ R/resampledata.R
@@ -4,10 +4,7 @@
 #' (2011) by Laura Chihara and Tim Hesterberg.
 #'
 #' @section Data sets:
-#' Type \code{data(package="resampledata")} to view all datasets. Note the
-#' following datasets were not included due to sharing restrictions and can only
-#'  be accessed on the textbook's webpage: \code{Bangladesh}, \code{GSS2002},
-#'  \code{GSS2006}, and \code{Volleyball2009}.
+#' Type \code{data(package="resampledata")} to view all datasets.
 #'
 #' @docType package
 #' @examples