+++ NEWS.md
@@ -1,6 +1,8 @@
 # resampledata 0.1.0.9000
 
 * Development version of package
+* Added originally excluded Bangladesh, GSS2002, GSS2006 and Volleyball2009 data
+sets.
 
 # resampledata 0.1.0